+++ _infra/docker/Dockerfile
@@ -1,4 +1,4 @@
-FROM openjdk:11-jre-slim
+FROM openjdk:17-slim
 
 RUN apt-get update
 COPY target/collectionexercisesvc.jar /opt/collectionexercisesvc.jar
